Entities Id's INT -> LONG

+++ app/src/main/java/com/example/gradeprojectdb01/data/entities/Instrument.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity
 data class Instrument (
     @PrimaryKey (autoGenerate = true)
-    val instrumentId: Int,
+    val instrumentId: Long,
     val name: String
 )
